Refactor: #77 - 키보드 extension 생성
