+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/Model/DFTextViewModel.swift
@@ -38,7 +38,7 @@ extension DFTextView{
             case (.trailing, .left): return .center
             case (.leading, .left): return .leading // 유지
             case (.trailing, .right): return .trailing // 유지
-//            default: return .center
+                //            default: return .center
         }
     }
 }
