+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/Frame/DesignFrame/DFModifyView.swift
@@ -123,7 +123,7 @@ private extension DFModifyView {
                 } label: {
                     Text("나가기")
                 }
-
+                
             } message: {
                 Text("종료 시 편집된 내용은 저장되지 않습니다.")
             }
