+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/Frame/DesignFrame/DFTextView.swift
@@ -14,9 +14,9 @@ struct DFTextView: View {
     let colorChip: [Color] = ColorPreset.colorPallete
     @Environment(\.displayScale) var displayScale
     @EnvironmentObject var imageModel: ImageListModel
-
+    
     @State var keyboardHeight: CGFloat = 0 // 키보드 높이 상태
-
+    
     var body: some View {
         VStack {
             Spacer()
@@ -51,76 +51,28 @@ struct DFTextView: View {
                 .onTapGesture {
                     isKeyboardVisible.toggle()
                 }
-
+            
             if tab == 0 {
                 fontSelector
-
+                
             } else if tab == 1 {
                 colorSelector
             }
-
+            
             textTabBar
             Spacer()
                 .frame(height: keyboardHeight)
-//                .padding(.bottom, keyboardHeight) // 키보드 높이만큼 여백 추가
         }
-//        .offset(y: -keyboardHeight / 2)
-                .animation(.easeOut(duration: 0.3))
-                
-                .keyboardHeight($keyboardHeight)
-//                .keyboardHeight($keyboardHeight)
-//        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .animation(.easeOut(duration: 0.3))
+        .keyboardHeight($keyboardHeight)
         .background(
             Color.black.opacity(0.5) // 반투명 검정색
         )
         .ignoresSafeArea(.keyboard)
         .onAppear {
             isKeyboardVisible = true // 뷰가 나타날 때 키보드 열기
-//            addKeyboardObservers() // 키보드 관찰자 추가
-        }
-        .onDisappear {
-//            removeKeyboardObservers() // 키보드 관찰자 제거
         }
         
     }
-
-}
-import Foundation
-import SwiftUI
-
-struct KeyboardProvider : ViewModifier {
     
-    //키보드 높이값
-    var keyboardHeight: Binding<CGFloat>
-    
-    func body(content: Content) -> some View {
-        content
-        //키보드 올라가기 직전 노티를 받으면 나오는 객체
-            .onReceive(NotificationCenter.default.publisher(for: UIResponder.keyboardWillShowNotification),
-                       perform: { notification in
-                guard let userInfo = notification.userInfo,
-                      let keyboardRect = userInfo[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect else { return }
-                
-                //키보드 높이값 . 바인딩 원본 객체 연결 -> 전달
-                self.keyboardHeight.wrappedValue = keyboardRect.height
-                
-            })
-        //키보드 닫기 전 보내는 노티 받으면 실행
-            .onReceive(NotificationCenter.default.publisher(for: UIResponder.keyboardWillHideNotification),
-                         perform: { _ in
-                //키보드 높이값 0으로 변경
-                self.keyboardHeight.wrappedValue = 0
-            })
-    }
-}
-
-public extension View {
-    func keyboardHeight(_ state: Binding<CGFloat>) -> some View {
-        self.modifier(KeyboardProvider(keyboardHeight: state))
-    }
-    
-    func hideKeyboard() {
-        let resign = #selector(UIResponder.resignFirstResponder)
-        UIApplication.shared.sendAction(resign, to: nil, from: nil, for: nil)
-      }
 }
